tasks/task01/python_project/requirements.txt
pyyaml
parameterized